Expanded prompt library to help workflow
Rather than writing prompts on the fly, a few prompts that are
important to the workflow have been pre-written in a json file.
This is greatly reducing the cognitive load of
having the prompts kept only in mind or in notes that require
searching.  This seems to also encourage thinking of other useful
prompts and refining their quality before using them.