Set landing page to demonstrate openAI connection
In the meantime, the catalog is auth guarded, so rather than
the landing page being just a sign in prompt, I wanted it to
look like the status check page to indicate that these things
are hooked up, and then a link that leads to the catalog from there.

I prefer that there's a catalog page that unauthenticated can view,
but currently it's mixed in with the librarian console dashboard.